# packages/contracts/openapi/travel-agent-v1.yaml
...
openapi: 3.1.1
info:
  title: 企业级智能旅行助手 API
  version: 1.0.0
  description: 浏览器仅通过 API Server 访问；会话使用 HttpOnly Cookie，状态变更请求使用 Double Submit CSRF。
servers:
  - url: http://localhost:8000
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialLoginRequest'
      responses:
        '200': { $ref: '#/components/responses/Accepted' }
        '401': { $ref: '#/components/responses/Error' }
      x-csrf-required: false
  /api/v1/auth/logout:
    post:
      operationId: logout
      security: [{ cookieAuth: [] }]
      responses:
        '204': { description: 会话已失效 }
        '401': { $ref: '#/components/responses/Error' }
      x-csrf-required: true
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: travel_agent_session
      description: HttpOnly 会话 Cookie，浏览器脚本不可读取。
...
  schemas:
    CredentialLoginRequest:
      type: object
      additionalProperties: false
      required: [account, password]
      properties:
        account: { type: string, pattern: '^[A-Za-z0-9_.@-]{3,64}$' }
        password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, trace_id, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code: { type: string, pattern: '^[a-z0-9_]{1,100}$' }
            message: { type: string, maxLength: 500 }
            retryable: { type: boolean }
        trace_id: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' }
        request_id: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' }
    AcceptedResult:
      type: object
      additionalProperties: true
  responses:
    Accepted:
      description: 请求已受理或资源查询成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AcceptedResult' }
    Error:
      description: 安全错误信封
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }